[Feature][Refactor Device Class Structure and Update ID Types][Organize Device Structure and Update ID Types for Supplier and Employee]

+++ ActionFactory/src/main/java/com/IntegrativeProject/ActionFactory/service/EmployeeService.java
@@ -26,7 +26,7 @@ public class EmployeeService {
         return this.employeeRepository.findAll();
     }
 
-    public  void  deleteEmployee(int id){
+    public  void  deleteEmployee(Long id){
         Optional<Employee> employee= this.employeeRepository.findById(id);
         if(employee.isPresent()){
             this.employeeRepository.delete(employee.get());
